fix: Fixed the Changing Legend Issue
