+++ src/components/RasterLayer.svelte
@@ -6,19 +6,15 @@
   import Accordion, { Panel } from '@smui-extra/accordion'
   import Fa from 'svelte-fa'
   import { faCalculator } from '@fortawesome/free-solid-svg-icons/faCalculator'
-  import { faRetweet } from '@fortawesome/free-solid-svg-icons/faRetweet'
   import { faDroplet } from '@fortawesome/free-solid-svg-icons/faDroplet'
   import { faList } from '@fortawesome/free-solid-svg-icons/faList'
   import Tab, { Label } from '@smui/tab'
   import TabBar from '@smui/tab-bar'
-  import Tooltip, { Wrapper } from '@smui/tooltip'
-
-  import Legend from './Legend.svelte'
-  import IntervalsLegend from './IntervalsLegend.svelte'
+  import GodLegend from './GodLegend.svelte'
   import { layerList, map } from '../stores'
   import type { Layer } from '../lib/types'
   import type { LayerSpecification } from '@maplibre/maplibre-gl-style-spec/types'
-  import { LayerInitialValues, DynamicLayerLegendTypes, DEFAULT_COLORMAP, TabNames } from '../lib/constants'
+  import { LayerInitialValues, DEFAULT_COLORMAP, TabNames } from '../lib/constants'
   import LayerNameGroup from './control-groups/LayerNameGroup.svelte'
   import OpacityPanel from './controls/OpacityPanel.svelte'
 
@@ -40,7 +36,6 @@
   let scalingValueStart = Math.floor(+layerBandMetadataMin * 10) / 10
   let scalingValueEnd = Math.ceil(+layerBandMetadataMax * 10) / 10
   let timer: ReturnType<typeof setTimeout>
-  let selectedLegendType = DynamicLayerLegendTypes.CONTINUOUS.toString()
 
   $: panelOpen, setLayerState()
   $: scalingValueStart, setScalingValueRange()
@@ -168,64 +163,7 @@
             <div class="action">
               <div class="content">
                 <div class="scene">
-                  <div class={`card ${selectedLegendType === DynamicLayerLegendTypes.INTERVALS ? 'is-flipped' : ''}`}>
-                    <div class="card-face card-face-front">
-                      <div class="container">
-                        <div class="flip">
-                          <Wrapper>
-                            <div
-                              style="cursor: pointer; width: 20px;  margin-left: auto;"
-                              on:click={() => {
-                                isLegendSwitchAnimate = true
-                                setTimeout(() => {
-                                  isLegendSwitchAnimate = false
-                                }, 400)
-
-                                selectedLegendType === DynamicLayerLegendTypes.INTERVALS
-                                  ? (selectedLegendType = DynamicLayerLegendTypes.CONTINUOUS)
-                                  : (selectedLegendType = DynamicLayerLegendTypes.INTERVALS)
-                              }}>
-                              <Fa icon={faRetweet} size="1x" spin={isLegendSwitchAnimate} />
-                            </div>
-                            <Tooltip showDelay={300} hideDelay={100} yPos="above">
-                              {selectedLegendType === DynamicLayerLegendTypes.INTERVALS
-                                ? `Show ${DynamicLayerLegendTypes.CONTINUOUS} legend`
-                                : `Show ${DynamicLayerLegendTypes.INTERVALS} legend`}
-                            </Tooltip>
-                          </Wrapper>
-                        </div>
-                        <Legend bind:activeColorMapName layerConfig={layer} />
-                      </div>
-                    </div>
-                    <div class="card-face card-face-back">
-                      <div class="container">
-                        <div class="flip">
-                          <Wrapper>
-                            <div
-                              style="cursor: pointer; width: 20px;  margin-left: auto;"
-                              on:click={() => {
-                                isLegendSwitchAnimate = true
-                                setTimeout(() => {
-                                  isLegendSwitchAnimate = false
-                                }, 400)
-
-                                selectedLegendType === DynamicLayerLegendTypes.INTERVALS
-                                  ? (selectedLegendType = DynamicLayerLegendTypes.CONTINUOUS)
-                                  : (selectedLegendType = DynamicLayerLegendTypes.INTERVALS)
-                              }}>
-                              <Fa icon={faRetweet} size="1x" spin={isLegendSwitchAnimate} />
-                            </div>
-                            <Tooltip showDelay={300} hideDelay={100} yPos="above">
-                              {selectedLegendType === DynamicLayerLegendTypes.INTERVALS
-                                ? `Show ${DynamicLayerLegendTypes.CONTINUOUS} legend`
-                                : `Show ${DynamicLayerLegendTypes.INTERVALS} legend`}
-                            </Tooltip>
-                          </Wrapper>
-                        </div>
-                        <IntervalsLegend bind:activeColorMapName layerConfig={layer} />
-                      </div>
-                    </div>
-                  </div>
+                  <GodLegend bind:activeColorMapName bind:layer />
                 </div>
               </div>
             </div>
